prevent user name enumeration in password reset
DokuWiki tries to be user friendly by telling users if they entered a
correct user name in the reset password form. Some consider this a
security problem as it allows for user name enumeration.

This adds an option to the plugin to supress all user hints in this work
flow.

Previously discussed in

* splitbrain/dokuwiki#1727
* splitbrain/dokuwiki#2928
* splitbrain/dokuwiki#2935
* splitbrain/dokuwiki#3540